+++ src/Application/inc/Application.hpp
@@ -39,7 +39,9 @@ namespace easyGUI
  * @brief Emulates a windowed application
  * 
  * @details The class acts as a wrapper around the ::sf::RenderWindow, and provides
- * easier management of the events (such as onClick, onKeyPress, etc.) 
+ * easier management of the events (such as onClick, onKeyPress, etc.). Furthermore,
+ * the window works with entire Menus of components, so that UI creation is faster
+ * and easier. 
  * 
  */
 #if defined(_WIN32) && BUILD_SHARED_LIBRARIES
@@ -154,14 +156,6 @@ public:
      * @details Closes the application window.
      */
     void stop();
-
-    // ----- Inherited from anchor -----
-
-    Point getLEFT() const override;
-    Point getRIGHT() const override;
-    Point getBOTTOM() const override;
-    Point getTOP() const override;
-    Point getCENTER() const override;
 private:
     static ::std::shared_ptr<Application> _instance;