+++ src/Application/inc/Exceptions/ApplicationException.hpp
@@ -29,15 +29,15 @@
 #include <string>
 
 #ifndef ERROR
-#define ERROR std::cout << "[ERROR] ";
+#define ERROR std::cout << "[ERROR] "
 #endif
 
 #ifndef WARN
-#define WARN std::cout << "[WARNING] ";
+#define WARN std::cout << "[WARNING] "
 #endif
 
 #ifndef INFO
-#define INFO std::cout << "[INFO] ";
+#define INFO std::cout << "[INFO] "
 #endif
 
 namespace easyGUI
@@ -64,8 +64,8 @@ public:
      * 
      * @param message The message to be displayed
      */
-    explicit ApplicationException(::std::string message) :
-        _msg("[APPLICATION] " + message + "\n");
+    explicit ApplicationException(const ::std::string& message) :
+        _msg("[APPLICATION] " + message + "\n") {}
 
     /**
      * @brief Returns the exception message